+++ test/gg_tree_expressions_test.dart
@@ -4,15 +4,15 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-import 'package:gg_template_project/gg_template_project.dart';
+import 'package:gg_tree_expressions/gg_tree_expressions.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('GgTemplateProject()', () {
+  group('GgTreeExpressions()', () {
     group('foo()', () {
       test('should return foo', () async {
-        const ggTemplateProject = GgTemplateProject();
-        expect(ggTemplateProject.foo(), 'foo');
+        const ggTreeExpressions = GgTreeExpressions();
+        expect(ggTreeExpressions.foo(), 'foo');
       });
     });
   });